chore: convert rest of jest tests
